Refactor and rename API-related stuff
